Initial import
First working version of the service.  It distributes access rights based on
which product groups were not allowed to be shown.

There is some logic necessary in other services because the current solution
only emits the data for the given delivery point's business-entity _and_ only if
the business-entity has a constraint.  For other cases the data is not
expressed.  An ideal solution for integrations would express all this as data
but we're limiting data creation for now.